+++ packages/adminui-vue/components/UTableEntity/store.js
@@ -12,6 +12,9 @@ const { Notification: $notify } = require('element-ui')
  * @param {number} pageSize Page size
  * @param {string} dateFormat Date format
  * @param {string} dateTimeFormat DateTime format
+ * @param {function(object):object} buildCopyConfig Function must return config for action "copy"
+ * @param {function(object):object} buildEditConfig Function must return config for action "edit"
+ * @param {function(object):object} buildAddNewConfig Function must return config for action "addNew"
  * @returns {Vuex} Store config
  */
 module.exports = ({
@@ -20,7 +23,10 @@ module.exports = ({
   columns,
   pageSize,
   dateFormat,
-  dateTimeFormat
+  dateTimeFormat,
+  buildCopyConfig,
+  buildEditConfig,
+  buildAddNewConfig
 }) => ({
   state () {
     return {
@@ -254,14 +260,14 @@ module.exports = ({
         entity: getters.entityName,
         formCode: getters.formCode
       })
-
-      $App.doCommand({
+      const config = buildAddNewConfig({
         cmdType: 'showForm',
         entity: getters.entityName,
         formCode: getters.formCode,
         target: $App.viewport.centralPanel,
         tabId
       })
+      $App.doCommand(config)
     },
 
     editRecord ({ getters }, ID) {
@@ -270,8 +276,7 @@ module.exports = ({
         instanceID: ID,
         formCode: getters.formCode
       })
-
-      $App.doCommand({
+      const config = buildEditConfig({
         cmdType: 'showForm',
         entity: getters.entityName,
         formCode: getters.formCode,
@@ -279,6 +284,7 @@ module.exports = ({
         target: $App.viewport.centralPanel,
         tabId
       })
+      $App.doCommand(config)
     },
 
     async deleteRecord ({ getters }, ID) {
@@ -304,8 +310,7 @@ module.exports = ({
         instanceID: ID,
         formCode: getters.formCode
       })
-
-      $App.doCommand({
+      const config = buildCopyConfig({
         cmdType: 'showForm',
         addByCurrent: true,
         entity: getters.entityName,
@@ -314,6 +319,7 @@ module.exports = ({
         target: $App.viewport.centralPanel,
         tabId
       })
+      $App.doCommand(config)
     },
 
     createLink ({ getters }, ID) {